風呂も判定する
